move CElement files
