+++ Common/src/geometry/primitives/meson.build
@@ -1,0 +1,7 @@
+common_src += files(['CElement.cpp',
+                     'CTRIA1.cpp',
+                     'CQUAD4.cpp',
+                     'CTETRA1.cpp',
+                     'CPYRAM5.cpp',
+                     'CPRISM6.cpp',
+                     'CHEXA8.cpp'])
